make line chart generic
